+++ internal/provider/plural_data_sources.go
@@ -101,6 +101,7 @@
 //go:generate go run generators/plural-data-source/main.go -data-source awscc_detective_member_invitations -cftype AWS::Detective::MemberInvitation -package detective ../aws/detective/member_invitation_plural_data_source_gen.go ../aws/detective/member_invitation_plural_data_source_gen_test.go
 //go:generate go run generators/plural-data-source/main.go -data-source awscc_devopsguru_notification_channels -cftype AWS::DevOpsGuru::NotificationChannel -package devopsguru ../aws/devopsguru/notification_channel_plural_data_source_gen.go ../aws/devopsguru/notification_channel_plural_data_source_gen_test.go
 //go:generate go run generators/plural-data-source/main.go -data-source awscc_directoryservice_simple_ads -cftype AWS::DirectoryService::SimpleAD -package directoryservice ../aws/directoryservice/simple_ad_plural_data_source_gen.go ../aws/directoryservice/simple_ad_plural_data_source_gen_test.go
+//go:generate go run generators/plural-data-source/main.go -data-source awscc_docdbelastic_clusters -cftype AWS::DocDBElastic::Cluster -package docdbelastic ../aws/docdbelastic/cluster_plural_data_source_gen.go ../aws/docdbelastic/cluster_plural_data_source_gen_test.go
 //go:generate go run generators/plural-data-source/main.go -data-source awscc_dynamodb_global_tables -cftype AWS::DynamoDB::GlobalTable -package dynamodb ../aws/dynamodb/global_table_plural_data_source_gen.go ../aws/dynamodb/global_table_plural_data_source_gen_test.go
 //go:generate go run generators/plural-data-source/main.go -data-source awscc_dynamodb_tables -cftype AWS::DynamoDB::Table -package dynamodb ../aws/dynamodb/table_plural_data_source_gen.go ../aws/dynamodb/table_plural_data_source_gen_test.go
 //go:generate go run generators/plural-data-source/main.go -data-source awscc_ec2_capacity_reservations -cftype AWS::EC2::CapacityReservation -package ec2 ../aws/ec2/capacity_reservation_plural_data_source_gen.go ../aws/ec2/capacity_reservation_plural_data_source_gen_test.go
@@ -501,6 +502,7 @@ import (
 	_ "github.com/hashicorp/terraform-provider-awscc/internal/aws/detective"
 	_ "github.com/hashicorp/terraform-provider-awscc/internal/aws/devopsguru"
 	_ "github.com/hashicorp/terraform-provider-awscc/internal/aws/directoryservice"
+	_ "github.com/hashicorp/terraform-provider-awscc/internal/aws/docdbelastic"
 	_ "github.com/hashicorp/terraform-provider-awscc/internal/aws/dynamodb"
 	_ "github.com/hashicorp/terraform-provider-awscc/internal/aws/ec2"
 	_ "github.com/hashicorp/terraform-provider-awscc/internal/aws/ecr"